feat: centralize data initialization
