+++ babytracker-pro/src/app/checklist/page.tsx
@@ -97,15 +97,7 @@ export default function EnhancedChecklistPage() {
   // Get current baby from store
   const { currentBaby, initializeData, initializeProfile } = useBabyTrackerStore()
 
-  useEffect(() => {
-    // Check for stored email and initialize profile from database
-    const storedEmail = typeof window !== 'undefined' ? localStorage.getItem('user-email') : null
-    if (storedEmail && storedEmail !== 'nouveau.utilisateur@example.com') {
-      console.log('Checklist page: initializing profile for email:', storedEmail)
-      initializeProfile(storedEmail)
-    }
-    initializeData()
-  }, [initializeData, initializeProfile])
+
 
   const ageInWeeks = currentBaby ? getAgeInWeeks(currentBaby.birthDate) : 0
