+++ babytracker-pro/src/app/diaper/page.tsx
@@ -107,15 +107,7 @@ export default function DiaperPage() {
   })
 
 
-  useEffect(() => {
-    // Check for stored email and initialize profile from database
-    const storedEmail = typeof window !== 'undefined' ? localStorage.getItem('user-email') : null
-    if (storedEmail && storedEmail !== 'nouveau.utilisateur@example.com') {
-      console.log('Diaper page: initializing profile for email:', storedEmail)
-      initializeProfile(storedEmail)
-    }
-    initializeData()
-  }, [initializeData, initializeProfile])
+
 
   const ageInWeeks = currentBaby ? getAgeInWeeks(currentBaby.birthDate) : 0
