+++ babytracker-pro/src/app/feeding/page.tsx
@@ -308,15 +308,6 @@ export default function FeedingPage() {
   })
 
   // ✅ TOUS les useEffect d'abord
-  useEffect(() => {
-    // Check for stored email and initialize profile from database
-    const storedEmail = typeof window !== 'undefined' ? localStorage.getItem('user-email') : null
-    if (storedEmail && storedEmail !== 'nouveau.utilisateur@example.com') {
-      console.log('Feeding page: initializing profile for email:', storedEmail)
-      initializeProfile(storedEmail)
-    }
-    initializeData()
-  }, [initializeData, initializeProfile])
 
   useEffect(() => {
     let interval: NodeJS.Timeout
