+++ babytracker-pro/src/app/growth/page.tsx
@@ -638,15 +638,6 @@ export default function AdvancedGrowthPage() {
   } = useBabyTrackerStore()
 
   // ✅ TOUS les useEffect
-  useEffect(() => {
-    // Check for stored email and initialize profile from database
-    const storedEmail = typeof window !== 'undefined' ? localStorage.getItem('user-email') : null
-    if (storedEmail && storedEmail !== 'nouveau.utilisateur@example.com') {
-      console.log('Growth page: initializing profile for email:', storedEmail)
-      initializeProfile(storedEmail)
-    }
-    initializeData()
-  }, [initializeData, initializeProfile])
 
   // ✅ Load milestones from database
   // ✅ Enhanced milestones loading with error handling
