+++ babytracker-pro/src/app/sleep/page.tsx
@@ -537,15 +537,7 @@ export default function SleepPage() {
   })
 
   // Initialize data on mount
-  useEffect(() => {
-    // Check for stored email and initialize profile from database
-    const storedEmail = typeof window !== 'undefined' ? localStorage.getItem('user-email') : null
-    if (storedEmail && storedEmail !== 'nouveau.utilisateur@example.com') {
-      console.log('Sleep page: initializing profile for email:', storedEmail)
-      initializeProfile(storedEmail)
-    }
-    initializeData()
-  }, [initializeData, initializeProfile])
+
 
   // Real-time timer effect
   useEffect(() => {
